fix: 创建报价返回quoteId + itemName自动填充 + lineNumber空值保护
- CpqQuoteController: 返回类型 R<Void> → R<Long>，返回新增报价ID
- CpqQuoteLineItemServiceImpl: 行项目创建时自动从产品填充itemName
- CpqQuoteLineItemMapper: selectMaxLineNumber增加IFNULL保护
- ICpqQuoteService/CpqQuoteServiceImpl: add方法返回值改为Long

+++ ruoyi-modules/ruoyi-cpq-quote/src/main/java/org/dromara/cpq/quote/service/ICpqQuoteService.java
@@ -13,7 +13,7 @@ public interface ICpqQuoteService extends IService<CpqQuote> {
     CpqQuoteVo selectById(Long id);
     List<CpqQuoteVo> selectList(CpqQuoteBo bo);
     TableDataInfo<CpqQuoteVo> selectPageList(CpqQuoteBo bo, PageQuery pageQuery);
-    int insert(CpqQuoteBo bo);
+    Long insert(CpqQuoteBo bo);
     int update(CpqQuoteBo bo);
     int deleteById(Long id);
     int deleteByIds(Long[] ids);
